Move review_assessment_details under review namespace
