+++ app/views/planning_applications/review/assessment_details/_information.html.erb
