+++ app/views/planning_applications/review/assessment_details/show.html.erb
