+++ app/views/planning_applications/review/tasks/index.html.erb
@@ -55,7 +55,7 @@
         ) %>
       <% end %>
       <%= render(
-        TaskListItems::ReviewAssessmentDetailsComponent.new(
+        TaskListItems::Review::AssessmentDetailsComponent.new(
           planning_application: @planning_application,
         )
       )%>
